refactor(apis/mcpserver): move query RPCs from command to query controller
Move getOAuthGrantStatus and getOrgOAuthApp from McpServerCommandController
to McpServerQueryController — both are read-only operations with can_view
authorization that belong on the query surface, not the command surface.

Regenerated all stubs (Go, Java, TypeScript, Python) and SDK clients
across all four SDKs. SDK codegen correctly routes both RPCs through
the query client.

Made-with: Cursor
