+++ apis/stubs/java/src/main/java/ai/stigmer/agentic/mcpserver/v1/McpServerQueryControllerGrpc.java
@@ -82,6 +82,68 @@ public final class McpServerQueryControllerGrpc {
     return getGetByReferenceMethod;
   }
 
+  private static volatile io.grpc.MethodDescriptor<ai.stigmer.agentic.mcpserver.v1.GetOAuthGrantStatusInput,
+      ai.stigmer.agentic.mcpserver.v1.GetOAuthGrantStatusOutput> getGetOAuthGrantStatusMethod;
+
+  @io.grpc.stub.annotations.RpcMethod(
+      fullMethodName = SERVICE_NAME + '/' + "getOAuthGrantStatus",
+      requestType = ai.stigmer.agentic.mcpserver.v1.GetOAuthGrantStatusInput.class,
+      responseType = ai.stigmer.agentic.mcpserver.v1.GetOAuthGrantStatusOutput.class,
+      methodType = io.grpc.MethodDescriptor.MethodType.UNARY)
+  public static io.grpc.MethodDescriptor<ai.stigmer.agentic.mcpserver.v1.GetOAuthGrantStatusInput,
+      ai.stigmer.agentic.mcpserver.v1.GetOAuthGrantStatusOutput> getGetOAuthGrantStatusMethod() {
+    io.grpc.MethodDescriptor<ai.stigmer.agentic.mcpserver.v1.GetOAuthGrantStatusInput, ai.stigmer.agentic.mcpserver.v1.GetOAuthGrantStatusOutput> getGetOAuthGrantStatusMethod;
+    if ((getGetOAuthGrantStatusMethod = McpServerQueryControllerGrpc.getGetOAuthGrantStatusMethod) == null) {
+      synchronized (McpServerQueryControllerGrpc.class) {
+        if ((getGetOAuthGrantStatusMethod = McpServerQueryControllerGrpc.getGetOAuthGrantStatusMethod) == null) {
+          McpServerQueryControllerGrpc.getGetOAuthGrantStatusMethod = getGetOAuthGrantStatusMethod =
+              io.grpc.MethodDescriptor.<ai.stigmer.agentic.mcpserver.v1.GetOAuthGrantStatusInput, ai.stigmer.agentic.mcpserver.v1.GetOAuthGrantStatusOutput>newBuilder()
+              .setType(io.grpc.MethodDescriptor.MethodType.UNARY)
+              .setFullMethodName(generateFullMethodName(SERVICE_NAME, "getOAuthGrantStatus"))
+              .setSampledToLocalTracing(true)
+              .setRequestMarshaller(io.grpc.protobuf.ProtoUtils.marshaller(
+                  ai.stigmer.agentic.mcpserver.v1.GetOAuthGrantStatusInput.getDefaultInstance()))
+              .setResponseMarshaller(io.grpc.protobuf.ProtoUtils.marshaller(
+                  ai.stigmer.agentic.mcpserver.v1.GetOAuthGrantStatusOutput.getDefaultInstance()))
+              .setSchemaDescriptor(new McpServerQueryControllerMethodDescriptorSupplier("getOAuthGrantStatus"))
+              .build();
+        }
+      }
+    }
+    return getGetOAuthGrantStatusMethod;
+  }
+
+  private static volatile io.grpc.MethodDescriptor<ai.stigmer.agentic.mcpserver.v1.GetOrgOAuthAppInput,
+      ai.stigmer.agentic.mcpserver.v1.GetOrgOAuthAppOutput> getGetOrgOAuthAppMethod;
+
+  @io.grpc.stub.annotations.RpcMethod(
+      fullMethodName = SERVICE_NAME + '/' + "getOrgOAuthApp",
+      requestType = ai.stigmer.agentic.mcpserver.v1.GetOrgOAuthAppInput.class,
+      responseType = ai.stigmer.agentic.mcpserver.v1.GetOrgOAuthAppOutput.class,
+      methodType = io.grpc.MethodDescriptor.MethodType.UNARY)
+  public static io.grpc.MethodDescriptor<ai.stigmer.agentic.mcpserver.v1.GetOrgOAuthAppInput,
+      ai.stigmer.agentic.mcpserver.v1.GetOrgOAuthAppOutput> getGetOrgOAuthAppMethod() {
+    io.grpc.MethodDescriptor<ai.stigmer.agentic.mcpserver.v1.GetOrgOAuthAppInput, ai.stigmer.agentic.mcpserver.v1.GetOrgOAuthAppOutput> getGetOrgOAuthAppMethod;
+    if ((getGetOrgOAuthAppMethod = McpServerQueryControllerGrpc.getGetOrgOAuthAppMethod) == null) {
+      synchronized (McpServerQueryControllerGrpc.class) {
+        if ((getGetOrgOAuthAppMethod = McpServerQueryControllerGrpc.getGetOrgOAuthAppMethod) == null) {
+          McpServerQueryControllerGrpc.getGetOrgOAuthAppMethod = getGetOrgOAuthAppMethod =
+              io.grpc.MethodDescriptor.<ai.stigmer.agentic.mcpserver.v1.GetOrgOAuthAppInput, ai.stigmer.agentic.mcpserver.v1.GetOrgOAuthAppOutput>newBuilder()
+              .setType(io.grpc.MethodDescriptor.MethodType.UNARY)
+              .setFullMethodName(generateFullMethodName(SERVICE_NAME, "getOrgOAuthApp"))
+              .setSampledToLocalTracing(true)
+              .setRequestMarshaller(io.grpc.protobuf.ProtoUtils.marshaller(
+                  ai.stigmer.agentic.mcpserver.v1.GetOrgOAuthAppInput.getDefaultInstance()))
+              .setResponseMarshaller(io.grpc.protobuf.ProtoUtils.marshaller(
+                  ai.stigmer.agentic.mcpserver.v1.GetOrgOAuthAppOutput.getDefaultInstance()))
+              .setSchemaDescriptor(new McpServerQueryControllerMethodDescriptorSupplier("getOrgOAuthApp"))
+              .build();
+        }
+      }
+    }
+    return getGetOrgOAuthAppMethod;
+  }
+
   /**
    * Creates a new async stub that supports all call types for the service
    */
@@ -183,6 +245,41 @@ public final class McpServerQueryControllerGrpc {
         io.grpc.stub.StreamObserver<ai.stigmer.agentic.mcpserver.v1.McpServer> responseObserver) {
       io.grpc.stub.ServerCalls.asyncUnimplementedUnaryCall(getGetByReferenceMethod(), responseObserver);
     }
+
+    /**
+     * <pre>
+     * Check whether the authenticated user has an active OAuth grant for
+     * an MCP server in the specified org.
+     * Returns grant metadata (connected status, token expiry, auth method)
+     * without exposing any secret token values. The frontend uses this to
+     * render the correct OAuth state in the MCP server detail page and
+     * session composer.
+     * &#64;internal
+     * Authorization: Requires can_view permission on the mcp_server resource.
+     * The resource_id field contains the MCP server's system-generated ID.
+     * </pre>
+     */
+    default void getOAuthGrantStatus(ai.stigmer.agentic.mcpserver.v1.GetOAuthGrantStatusInput request,
+        io.grpc.stub.StreamObserver<ai.stigmer.agentic.mcpserver.v1.GetOAuthGrantStatusOutput> responseObserver) {
+      io.grpc.stub.ServerCalls.asyncUnimplementedUnaryCall(getGetOAuthGrantStatusMethod(), responseObserver);
+    }
+
+    /**
+     * <pre>
+     * Query whether an org has a BYOA override for a resource.
+     * Returns override metadata (existence, OAuthApp ID, client_id) without
+     * exposing secrets. The frontend uses this to show which credential
+     * source is active and to offer override management options to org admins.
+     * &#64;internal
+     * Authorization: Requires can_view permission on the mcp_server resource.
+     * Any user who can view the MCP server can check whether their org has
+     * an override — no secrets are exposed.
+     * </pre>
+     */
+    default void getOrgOAuthApp(ai.stigmer.agentic.mcpserver.v1.GetOrgOAuthAppInput request,
+        io.grpc.stub.StreamObserver<ai.stigmer.agentic.mcpserver.v1.GetOrgOAuthAppOutput> responseObserver) {
+      io.grpc.stub.ServerCalls.asyncUnimplementedUnaryCall(getGetOrgOAuthAppMethod(), responseObserver);
+    }
   }
 
   /**
@@ -260,6 +357,43 @@ public final class McpServerQueryControllerGrpc {
       io.grpc.stub.ClientCalls.asyncUnaryCall(
           getChannel().newCall(getGetByReferenceMethod(), getCallOptions()), request, responseObserver);
     }
+
+    /**
+     * <pre>
+     * Check whether the authenticated user has an active OAuth grant for
+     * an MCP server in the specified org.
+     * Returns grant metadata (connected status, token expiry, auth method)
+     * without exposing any secret token values. The frontend uses this to
+     * render the correct OAuth state in the MCP server detail page and
+     * session composer.
+     * &#64;internal
+     * Authorization: Requires can_view permission on the mcp_server resource.
+     * The resource_id field contains the MCP server's system-generated ID.
+     * </pre>
+     */
+    public void getOAuthGrantStatus(ai.stigmer.agentic.mcpserver.v1.GetOAuthGrantStatusInput request,
+        io.grpc.stub.StreamObserver<ai.stigmer.agentic.mcpserver.v1.GetOAuthGrantStatusOutput> responseObserver) {
+      io.grpc.stub.ClientCalls.asyncUnaryCall(
+          getChannel().newCall(getGetOAuthGrantStatusMethod(), getCallOptions()), request, responseObserver);
+    }
+
+    /**
+     * <pre>
+     * Query whether an org has a BYOA override for a resource.
+     * Returns override metadata (existence, OAuthApp ID, client_id) without
+     * exposing secrets. The frontend uses this to show which credential
+     * source is active and to offer override management options to org admins.
+     * &#64;internal
+     * Authorization: Requires can_view permission on the mcp_server resource.
+     * Any user who can view the MCP server can check whether their org has
+     * an override — no secrets are exposed.
+     * </pre>
+     */
+    public void getOrgOAuthApp(ai.stigmer.agentic.mcpserver.v1.GetOrgOAuthAppInput request,
+        io.grpc.stub.StreamObserver<ai.stigmer.agentic.mcpserver.v1.GetOrgOAuthAppOutput> responseObserver) {
+      io.grpc.stub.ClientCalls.asyncUnaryCall(
+          getChannel().newCall(getGetOrgOAuthAppMethod(), getCallOptions()), request, responseObserver);
+    }
   }
 
   /**
@@ -316,6 +450,41 @@ public final class McpServerQueryControllerGrpc {
       return io.grpc.stub.ClientCalls.blockingV2UnaryCall(
           getChannel(), getGetByReferenceMethod(), getCallOptions(), request);
     }
+
+    /**
+     * <pre>
+     * Check whether the authenticated user has an active OAuth grant for
+     * an MCP server in the specified org.
+     * Returns grant metadata (connected status, token expiry, auth method)
+     * without exposing any secret token values. The frontend uses this to
+     * render the correct OAuth state in the MCP server detail page and
+     * session composer.
+     * &#64;internal
+     * Authorization: Requires can_view permission on the mcp_server resource.
+     * The resource_id field contains the MCP server's system-generated ID.
+     * </pre>
+     */
+    public ai.stigmer.agentic.mcpserver.v1.GetOAuthGrantStatusOutput getOAuthGrantStatus(ai.stigmer.agentic.mcpserver.v1.GetOAuthGrantStatusInput request) throws io.grpc.StatusException {
+      return io.grpc.stub.ClientCalls.blockingV2UnaryCall(
+          getChannel(), getGetOAuthGrantStatusMethod(), getCallOptions(), request);
+    }
+
+    /**
+     * <pre>
+     * Query whether an org has a BYOA override for a resource.
+     * Returns override metadata (existence, OAuthApp ID, client_id) without
+     * exposing secrets. The frontend uses this to show which credential
+     * source is active and to offer override management options to org admins.
+     * &#64;internal
+     * Authorization: Requires can_view permission on the mcp_server resource.
+     * Any user who can view the MCP server can check whether their org has
+     * an override — no secrets are exposed.
+     * </pre>
+     */
+    public ai.stigmer.agentic.mcpserver.v1.GetOrgOAuthAppOutput getOrgOAuthApp(ai.stigmer.agentic.mcpserver.v1.GetOrgOAuthAppInput request) throws io.grpc.StatusException {
+      return io.grpc.stub.ClientCalls.blockingV2UnaryCall(
+          getChannel(), getGetOrgOAuthAppMethod(), getCallOptions(), request);
+    }
   }
 
   /**
@@ -371,6 +540,41 @@ public final class McpServerQueryControllerGrpc {
     public ai.stigmer.agentic.mcpserver.v1.McpServer getByReference(ai.stigmer.commons.apiresource.ApiResourceReference request) {
       return io.grpc.stub.ClientCalls.blockingUnaryCall(
           getChannel(), getGetByReferenceMethod(), getCallOptions(), request);
+    }
+
+    /**
+     * <pre>
+     * Check whether the authenticated user has an active OAuth grant for
+     * an MCP server in the specified org.
+     * Returns grant metadata (connected status, token expiry, auth method)
+     * without exposing any secret token values. The frontend uses this to
+     * render the correct OAuth state in the MCP server detail page and
+     * session composer.
+     * &#64;internal
+     * Authorization: Requires can_view permission on the mcp_server resource.
+     * The resource_id field contains the MCP server's system-generated ID.
+     * </pre>
+     */
+    public ai.stigmer.agentic.mcpserver.v1.GetOAuthGrantStatusOutput getOAuthGrantStatus(ai.stigmer.agentic.mcpserver.v1.GetOAuthGrantStatusInput request) {
+      return io.grpc.stub.ClientCalls.blockingUnaryCall(
+          getChannel(), getGetOAuthGrantStatusMethod(), getCallOptions(), request);
+    }
+
+    /**
+     * <pre>
+     * Query whether an org has a BYOA override for a resource.
+     * Returns override metadata (existence, OAuthApp ID, client_id) without
+     * exposing secrets. The frontend uses this to show which credential
+     * source is active and to offer override management options to org admins.
+     * &#64;internal
+     * Authorization: Requires can_view permission on the mcp_server resource.
+     * Any user who can view the MCP server can check whether their org has
+     * an override — no secrets are exposed.
+     * </pre>
+     */
+    public ai.stigmer.agentic.mcpserver.v1.GetOrgOAuthAppOutput getOrgOAuthApp(ai.stigmer.agentic.mcpserver.v1.GetOrgOAuthAppInput request) {
+      return io.grpc.stub.ClientCalls.blockingUnaryCall(
+          getChannel(), getGetOrgOAuthAppMethod(), getCallOptions(), request);
     }
   }
 
@@ -430,10 +634,49 @@ public final class McpServerQueryControllerGrpc {
       return io.grpc.stub.ClientCalls.futureUnaryCall(
           getChannel().newCall(getGetByReferenceMethod(), getCallOptions()), request);
     }
+
+    /**
+     * <pre>
+     * Check whether the authenticated user has an active OAuth grant for
+     * an MCP server in the specified org.
+     * Returns grant metadata (connected status, token expiry, auth method)
+     * without exposing any secret token values. The frontend uses this to
+     * render the correct OAuth state in the MCP server detail page and
+     * session composer.
+     * &#64;internal
+     * Authorization: Requires can_view permission on the mcp_server resource.
+     * The resource_id field contains the MCP server's system-generated ID.
+     * </pre>
+     */
+    public com.google.common.util.concurrent.ListenableFuture<ai.stigmer.agentic.mcpserver.v1.GetOAuthGrantStatusOutput> getOAuthGrantStatus(
+        ai.stigmer.agentic.mcpserver.v1.GetOAuthGrantStatusInput request) {
+      return io.grpc.stub.ClientCalls.futureUnaryCall(
+          getChannel().newCall(getGetOAuthGrantStatusMethod(), getCallOptions()), request);
+    }
+
+    /**
+     * <pre>
+     * Query whether an org has a BYOA override for a resource.
+     * Returns override metadata (existence, OAuthApp ID, client_id) without
+     * exposing secrets. The frontend uses this to show which credential
+     * source is active and to offer override management options to org admins.
+     * &#64;internal
+     * Authorization: Requires can_view permission on the mcp_server resource.
+     * Any user who can view the MCP server can check whether their org has
+     * an override — no secrets are exposed.
+     * </pre>
+     */
+    public com.google.common.util.concurrent.ListenableFuture<ai.stigmer.agentic.mcpserver.v1.GetOrgOAuthAppOutput> getOrgOAuthApp(
+        ai.stigmer.agentic.mcpserver.v1.GetOrgOAuthAppInput request) {
+      return io.grpc.stub.ClientCalls.futureUnaryCall(
+          getChannel().newCall(getGetOrgOAuthAppMethod(), getCallOptions()), request);
+    }
   }
 
   private static final int METHODID_GET = 0;
   private static final int METHODID_GET_BY_REFERENCE = 1;
+  private static final int METHODID_GET_OAUTH_GRANT_STATUS = 2;
+  private static final int METHODID_GET_ORG_OAUTH_APP = 3;
 
   private static final class MethodHandlers<Req, Resp> implements
       io.grpc.stub.ServerCalls.UnaryMethod<Req, Resp>,
@@ -459,6 +702,14 @@ public final class McpServerQueryControllerGrpc {
         case METHODID_GET_BY_REFERENCE:
           serviceImpl.getByReference((ai.stigmer.commons.apiresource.ApiResourceReference) request,
               (io.grpc.stub.StreamObserver<ai.stigmer.agentic.mcpserver.v1.McpServer>) responseObserver);
+          break;
+        case METHODID_GET_OAUTH_GRANT_STATUS:
+          serviceImpl.getOAuthGrantStatus((ai.stigmer.agentic.mcpserver.v1.GetOAuthGrantStatusInput) request,
+              (io.grpc.stub.StreamObserver<ai.stigmer.agentic.mcpserver.v1.GetOAuthGrantStatusOutput>) responseObserver);
+          break;
+        case METHODID_GET_ORG_OAUTH_APP:
+          serviceImpl.getOrgOAuthApp((ai.stigmer.agentic.mcpserver.v1.GetOrgOAuthAppInput) request,
+              (io.grpc.stub.StreamObserver<ai.stigmer.agentic.mcpserver.v1.GetOrgOAuthAppOutput>) responseObserver);
           break;
         default:
           throw new AssertionError();
@@ -492,6 +743,20 @@ public final class McpServerQueryControllerGrpc {
               ai.stigmer.commons.apiresource.ApiResourceReference,
               ai.stigmer.agentic.mcpserver.v1.McpServer>(
                 service, METHODID_GET_BY_REFERENCE)))
+        .addMethod(
+          getGetOAuthGrantStatusMethod(),
+          io.grpc.stub.ServerCalls.asyncUnaryCall(
+            new MethodHandlers<
+              ai.stigmer.agentic.mcpserver.v1.GetOAuthGrantStatusInput,
+              ai.stigmer.agentic.mcpserver.v1.GetOAuthGrantStatusOutput>(
+                service, METHODID_GET_OAUTH_GRANT_STATUS)))
+        .addMethod(
+          getGetOrgOAuthAppMethod(),
+          io.grpc.stub.ServerCalls.asyncUnaryCall(
+            new MethodHandlers<
+              ai.stigmer.agentic.mcpserver.v1.GetOrgOAuthAppInput,
+              ai.stigmer.agentic.mcpserver.v1.GetOrgOAuthAppOutput>(
+                service, METHODID_GET_ORG_OAUTH_APP)))
         .build();
   }
 
@@ -542,6 +807,8 @@ public final class McpServerQueryControllerGrpc {
               .setSchemaDescriptor(new McpServerQueryControllerFileDescriptorSupplier())
               .addMethod(getGetMethod())
               .addMethod(getGetByReferenceMethod())
+              .addMethod(getGetOAuthGrantStatusMethod())
+              .addMethod(getGetOrgOAuthAppMethod())
               .build();
         }
       }
